backend/requirements.txt
# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0

# --- Database & Migrations ---
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
asyncpg==0.30.0
alembic==1.14.0
chromadb==0.5.21

# --- Security & Validation ---
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Web3
web3==6.11.1
py-solc-x

# --- AI & External APIs ---
groq==0.11.0
google-generativeai==0.8.3
httpx==0.27.2

# --- Utilities ---
python-dotenv==1.0.1
pillow==10.4.0
python-magic==0.4.27


celery==5.3.6
redis==5.0.1
exif==1.6.0


sentence-transformers==3.0.1
pydub==0.25.1

# Jwt and Google OAuth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.22.0